+++ src/components/Code/Editor.js
@@ -21,46 +21,18 @@ const EditorContainer = styled.div`
   }
 `;
 
-const javascriptInitialEditorState = `function square(x) {
-    // your code goes here
-  
-  };
-  `;
-
-const pythonInitialEditorState = `# example code: print n fibonacci numbers
-  
-  a = 10
-  
-  def fib(n):
-      a, b = 0, 1
-      for _ in range(n):
-          yield a
-          a, b = b, a + b
-  
-  print(list(fib(a)))
-  `;
-
 const input1 = 3;
 const input2 = 10;
 const input3 = 329425;
 
-const Editor = ({ output, setOutput }) => {
-  const [editorState, setEditorState] = React.useState();
-  const [language, setLanguage] = React.useState('javascript');
-
-  const handleSelection = event => {
-    setLanguage(event.target.value);
-    setEditorState(mapLanguageToEditorState(event.target.value));
-  };
-
-  const handlePost = () => {
-    const tests = [input1, input2, input3];
-    if (mapLanguageToId(language) === 63) {
-      tests.forEach(el => setOutput(testCode('square', el)));
-    } else {
-      setOutput(logCode(editorState, mapLanguageToId(language)));
-    }
-  };
+const Editor = ({
+  output,
+  setOutput,
+  language,
+  setLanguage,
+  editorState,
+  setEditorState,
+}) => {
 
   return (
     <EditorContainer>
@@ -69,7 +41,9 @@ const Editor = ({ output, setOutput }) => {
         value={editorState}
         options={{
           mode: `${
-            language === "java" || language === "cpp" ? "clike" : language
+            language === 'java' || language === 'cpp'
+              ? 'clike'
+              : language
           }`,
           theme: 'material',
           lineNumbers: true,
